Got my DistillerGRUCell producing output of the same dimension as the Torch GRUCell implementation. Changed the name of the test script.
